api/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.0
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.14.0
httptools==0.5.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.2
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
python-dotenv==0.21.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tiktoken==0.3.2
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==10.4
yarl==1.8.2
